Refactor controls into MatrixComponent, fix bug with subscriptions in CircleComponent
