+++ src/app/matrix/matrix.component.html
@@ -1,6 +1,6 @@
 <div class="matrix-flex" [ngStyle]="matrixContainer">
   <div *ngFor="let interval of matrix" class="interval" [ngStyle]="interval.color"
-       (mousedown)="clickPlay(interval)">
+       (mousedown)="clickPlay($event, interval)">
     <svg width="100%" height="100%">
       <text x="50%" y="50%" dominant-baseline="middle" text-anchor="middle">
         {{interval.value}}
