+++ src/app/synth.service.ts
@@ -1,8 +1,7 @@
 import { Injectable } from '@angular/core';
 import {Interval} from './interval-model';
-import {BehaviorSubject, fromEvent, merge, Observable, Subject} from 'rxjs';
+import {BehaviorSubject, Subject} from 'rxjs';
 import {Note} from './note.model';
-import {filter, first, map, take} from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root'
@@ -13,44 +12,14 @@ export class SynthService {
   analyser: AnalyserNode;
   note: Subject<Note> = new Subject<Note>();
   curve: BehaviorSubject<number[][]> = new BehaviorSubject<number[][]>([]);
-  keyDown: Observable<any> = fromEvent(document, 'keydown').pipe(
-    filter(((event: KeyboardEvent) => !this.pressedKeys.map(key => key.interval.key).includes(event.key))),
-    map((event: KeyboardEvent) => this.validKeys.indexOf(event.key)),
-    filter(index => index >= 0)
-  );
-  defaultKeys: string[] = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'q', 'w', 'e', 'r', 't', 'y', 'u',
-    'i', 'o', 'p', 'a', 's', 'd', 'f'];
-  validKeys: string[] = [];
-  pressedKeys: Note[] = [];
   amplitude: Subject<number> = new Subject<number>();
   constructor() {}
-  play(interval) {
-    const note = this.generateNote(interval);
-    if (interval.key) {
-      this.pressedKeys.push(note);
-      const keyup = fromEvent(document, 'keyup').pipe(
-        filter((key: KeyboardEvent) => this.validKeys.includes(key.key)),
-        first()
-      );
-      keyup.subscribe(next => {
-        note.releaseNote();
-        this.pressedKeys = this.pressedKeys.filter(key => key.interval.key !== interval.key).slice();
-      });
-    } else {
-      const up = fromEvent(event.target, 'mouseup').pipe(take(1));
-      const leave = fromEvent(event.target, 'mouseleave').pipe(take(1));
-      merge(up, leave).subscribe(a => {
-        note.releaseNote();
-      });
-    }
-    this.note.next(note);
-    this.draw();
-  }
   initialize() {
     this.context = new AudioContext();
     this.analyser = this.context.createAnalyser();
     this.analyser.connect(this.context.destination);
     this.initialized = true;
+    this.draw();
   }
   generateNote(interval: Interval): Note {
     if (!this.initialized) {
